libierty: change _ANSIDECL_H to ANSIDECL_H
ansidecl.h is using _ANSIDECL_H as it's guard #ifdef.  But names that start with
underscore are reserved, and the latest version of Xcode seems to be exercising
its prerogative to define _ANSIDECL_H.

This patch changes _ANSIDECL_H to ANSIDECL_H so Xcode can still be used to
bootstrap gcc.

Signed-off-by: Lawrence D'Anna <larry@elder-gods.org>